Usermanagement/Usermanagement/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
django-cors-headers==4.5.0
python-dotenv==1.0.1
pillow==10.3.0


# beautifulsoup4==4.12.3
# blinker==1.8.2
# blivet==3.10.0
# # blivet-gui==2.5.0
# boto3==1.35.63
# botocore==1.35.63
# Brlapi==0.8.5
# cached-property==1.5.2
# certifi==2023.5.7
# cffi==1.16.0
# chardet==3.0.4
# charset-normalizer==2.0.12
# click==8.1.7
# click-plugins==1.1.1
# cryptography==41.0.7
# cupshelpers==1.0
# dasbus==1.7
# dbus-python==1.3.2
# distro==1.9.0
# Django==5.1.2
# django-cors-headers==4.5.0
# django-mjml==1.3
# django-rest-framework==0.1.0
# djangorestframework==3.15.2
# djangorestframework-simplejwt==5.3.1
# dnf==4.21.1
# docker==6.1.3
# docker-compose==1.29.2
# docker-pycreds==0.4.0
# dockerpty==0.4.1
# docopt==0.6.2
# evdev==1.7.1
# fedora-third-party==0.10
# file-magic==0.4.0
# Flask==3.0.3
# Flask-Cors==5.0.0
# Flask-Migrate==4.0.7
# Flask-SQLAlchemy==3.1.1
# fros==1.1
# greenlet==3.1.1
# humanize==3.13.1
# idna==2.10
# iso639==0.1.4
# itsdangerous==2.2.0
# Jinja2==3.1.4
# jmespath==1.0.1
# jsonschema==4.19.1
# jsonschema-specifications==2023.11.2
# langtable==0.0.68
# libcomps==0.1.20
# libdnf==0.73.3
# louis==3.28.0
# lutris==0.5.17
# lxml==5.1.0
# Mako==1.2.3
# MarkupSafe==2.1.3
# moddb==0.11.0
# mysqlclient==2.0.1
# nftables==0.1
# olefile==0.47
# packaging==23.2
# paramiko==3.5.0
# Paste==3.7.1
# pexpect==4.9.0
# pid==2.2.3
# pika==1.1.0

# ply==3.11
# productmd==1.41
# progressbar2==3.53.2
# protobuf==3.19.6
# psycopg2-binary==2.9.10
# ptyprocess==0.7.0
# pwquality==1.4.5
# pyasn1==0.6.0
# pycairo==1.25.1
# pycparser==2.20
# pycrypto==2.6.1
# pycups==2.0.4
# pyenchant==3.2.2
# pygit2==1.14.0
# PyGObject==3.48.2
# PyJWT==2.10.1
# pykickstart==3.52
# PyNaCl==1.5.0
# pyOpenSSL==23.2.0
# pyparted==3.13.0
# pypresence==4.3.0
# PySocks==1.7.1
# python-augeas==1.1.0
# python-dateutil==2.8.2
# python-dotenv==1.0.1
# python-meh==0.51
# python-pam==2.0.2
# python-utils==3.7.0
# python-vlc==3.0.20123
# pyudev==0.24.1
# pyxdg==0.27
# PyYAML==6.0.1
# referencing==0.31.1
# regex==2024.9.11
# requests==2.26.0
# requests-file==2.0.0
# requests-ftp==0.3.1
# rpds-py==0.18.1
# rpm==4.19.1.1
# rpmautospec==0.7.3
# rpmautospec-core==0.1.5
# s3transfer==0.10.3
# # selinux @ file:///builddir/build/BUILD/libselinux-3.7/src
# # sepolicy @ file:///builddir/build/BUILD/selinux-3.7/python/sepolicy
# setools==4.5.1
# setuptools==75.2.0
# simpleaudio==1.0.4
# simpleline==1.9.0
# six==1.16.0
# sos==4.7.2
# soupsieve==2.5
# SQLAlchemy==2.0.36
# sqlparse==0.5.1
# systemd-python==235
# Tempita==0.5.2
# texttable==1.6.7
# typing_extensions==4.12.2
# urllib3==1.26.20
# websocket-client==1.3.3
# Werkzeug==3.0.4
# xkbregistry==0.3
